+++ src/lib.rs
@@ -1,13 +1,14 @@
+pub mod control_transfer;
 pub mod descriptors;
 pub mod os;
+#[cfg(target_os = "linux")]
+pub use control_transfer::ControlTransfer;
 #[cfg(target_os = "linux")]
 pub use os::linux::constants::*;
 #[cfg(target_os = "linux")]
 pub use os::linux::enumerate::UsbEnumerate;
 #[cfg(target_os = "linux")]
 pub use os::linux::usb_device::UsbDevice;
-#[cfg(target_os = "linux")]
-pub use os::linux::usbfs::ControlTransfer;
 #[cfg(target_os = "linux")]
 pub use os::linux::usbfs::UsbFs as UsbCore;
 #[cfg(target_os = "linux")]