+++ changelog.md
@@ -1,3 +1,8 @@
+# 0.3.0
+
+ - Change ControlTransfer API and setup an mmap fort control transfers.
+ - get_descriptor_string variants now return error on fails.
+
 # 0.2.4
 
  - Fix some of the unsafe memory leaks